refactor: per review(WIP)
