+++ src/meta-srv/src/lib.rs
@@ -24,6 +24,7 @@ pub mod cluster;
 pub mod election;
 pub mod error;
 mod failure_detector;
+pub mod flow_meta_alloc;
 pub mod handler;
 pub mod key;
 pub mod lease;
